Connect U4 +1V1 regulator input to +3V3, not +2V5
The FPGA can drive outputs unexpectedly during power supply ramp-up and
ramp-down. The +2V5 regulator starts abruptly when +3V3 approaches 2 V
and starts more slowly with +1V1 chained from it. With +1V1 chained from
+3V3 instead of from +2V5, both +1V1 and +2V5 supplies start up faster,
minimizing the duration of FPGA output glitches. This reduces the
duration of the glitch that causes #9.