Update CFG pull-up and pull-down resistor values
FPGA-TN-02039 says that 4.7k is recommended for pull-ups and that
pull-downs should be less than 500 ohms.

FPGA-TN-02038 says that pull-ups can be from 1k to 10k but is unclear on
pull-downs. Perhaps it is trying to suggest that pull-downs should be 0
ohms.

We've been using 10k for both pull-downs and pull-ups. We haven't
observed any configuration failures, but the pull-downs are clearly not
as effective as they could be, pulling the signals down to 1 V.

This changes pull-ups from 10k to 5.1k and pull-downs from 10k to 330
ohms.